+++ server/index.js
@@ -1,4 +1,4 @@
-import app from './api.js';
+import app from './api/api.js';
 
 const port = 3000;
